Provide non-empty diff/Changes files for four tests.
Four tests check if the Changes file supplied by a skeleton was
overridden in the package, but that did not generate the desired tag
when the file was removed from the skeleton, because the overriding
file was empty. A non-empty file is a better choice. This commit
provides one each for four tests.

Signed-off-by: Chris Lamb <lamby@debian.org>