Update copyright information for the "pedantic" template set.
After combining files from both template sets, the copyright for
'pedantic' was amended to also include the information for 'default'
as well. The latter may be dropped in the near future.

With the most recent entry going back more than nine years, an
additional entry was made for derivative work.

Signed-off-by: Chris Lamb <lamby@debian.org>