Remove code for executing post-template command from t/runtests.
The active templates implemented here do not work with the test runner
that is part of Lintian. The reasons are unknown; they could be
related to close() not working properly in a threaded environment.

In reaction to #916087, this code was removed. In the suite 'tests'
the watch file will now be supplied by an extra template.

Signed-off-by: Chris Lamb <lamby@debian.org>