Provide functionality to write configuration files.
Adds a function to Test::Lintian::ConfigFile that allows the writing
of configuration files.

It will be used to write all the values used to fill the templates to
the 'desc' in the test working directory in ./debian/test-out/. That
file is a great place to look when debugging tests whose templates
were filled with unexpected values.

Signed-off-by: Chris Lamb <lamby@debian.org>