Refer to test cases using paths inside the persistent runners.
This change further separates the test preparation and the test running
stages by referring to tests using directories.

A side effect is that the 'desc' file in the test working directory.
which holds data exchanged between the two stages, contains all values
used to fill the templates, including dynamically created values and
any defaults. That information can be valuable when debugging tests
(or builders).

Signed-off-by: Chris Lamb <lamby@debian.org>