Set $ENV{PERL_PATH_TINY_NO_FLOCK} when using IO::Async::Routine.
As demonstrated in #916087, files in the test runner are not always
closed properly. The issue is probably caused by an interference
between IO::Async::Routine, which implements the persistent runners,
to other forms of spawning processes.

This commit prevents the runners from hanging when using Path::Tiny.
Without this commit, Path::Tiny uses flock and may rely on close() to
remove the lock (which does not work).

This issue can be solved by migrating the tests to TAP::Harness. It is
known not to suffer from the same problem.

Signed-off-by: Chris Lamb <lamby@debian.org>