Use rewritten Lintian call for Perl coverage (also separates sorting).
This commit uses a rewritten call to Lintian. It draws on environment
variables to set up Perl coverage, if needed.

As an automatic consequence of this change, the sorting of tags now
takes place separately. Going forward, sorting only happens once after
the call to sed, if applicable, and not twice.

Signed-off-by: Chris Lamb <lamby@debian.org>