Remove diff directories and contents for two tests. (MR: !103)
The `./diff` directory exists to test specific features of patch
systems when upstream sources that that are unpackaged when building a
debian patch differ from contents in a nearby `orig.tar.gz` tarfile.
In the present cases, the contents of the `./diff` directories are
identical to that of `./orig`. Thus, there are no differences. Also,
neither test seems to exercise a potential false positive related to
the feature. Both `./diff` directories are removed.

Signed-off-by: Chris Lamb <lamby@debian.org>