Add description of 'updeb' to test description. (MR: !104)
The initial commit for this test from Fri Jun 8 17:29:00 2012 +0200
includes a helpful explanation of what is meant by 'updeb'. The
explanation is not present in the test files and could be lost if the
files ever become detached from the commit message, for example
because of a move. A brief explanation is added to the test
description.

Signed-off-by: Chris Lamb <lamby@debian.org>