Fix path to expected tags when those tags were calibrated.
Two tests use a "calibration" script that adjusts the expected tags.
In that particular circumstance, commit b81de6ce causes the displayed
diff to be incorrect (although the test result itself is reliable).
Here we fix the path if the test was calibrated. Now the diff is
correct again.

Possible long-term solutions are: (1) reversal of the "calibration"
logic in that it modifies the actual tags rather than the expected
tags; (2) a more powerful tag parts that compares only invariant
parts.

Signed-off-by: Chris Lamb <lamby@debian.org>