Require compat 11 for tests that needs it.
The option dh_clean -k was completely removed in compat 12.  The test
warning about its obsolescence requires compat 11 or lower. The
alternative would be to remove the test.

For another test, the command passing with '--' to dh_installdebs
fails. This likewise seems to be a new debhelper restriction. It also
requires compat 11.

Signed-off-by: Chris Lamb <lamby@debian.org>