Adjust to groff output using left single quote instead of apostrophe.
In its error output, groff-base 1.22.4-1 seems to delimit the left
side of macro names with a single quote instead of an apostrophe,
which was used until now.  Adjust the expected tags in the test
'manpage-general' and a line in check/manpages.pm that skips some such
lines.

It may be worthwhile to send a note to all maintainers of packages
that override such a tag. The change is hard to see and can be
frustrating to find.