Skip some input file information prepended by groff-base.
groff-base 1.22.4-1 prepends the name of the input file to any error
output. In Lintian, the messages look like '<standard input>:' for
regular man pages, and 'man/man3/XXX.3.gz:' if the error occurs in a
file included from elsewhere. For standard input, skip the information
to match the old tag output. For files included from elsewhere, the
extra information may be useful; it is retained.

If a file included from elsewhere causes an error but is also tested
on its own, the new behavior results in two similar tags being issued
that point out the same error (it would not be a duplicate).  Perhaps
the tag should not be skipped for included files.