Fix test runtests-calibration for repeat runs.
Previous versions appended to the input file, but it breaks repeated
runs. With adaptive rebuilding, the input file is not necessarily
generated on every run. It may already contain the appended line from
a previous run. The line is then duplicated. For this test, the
solution is to replace the file.

Signed-off-by: Chris Lamb <lamby@debian.org>