Fix debhelper-compat-experimental test after compat upgrade. (!MR 117)
This test was recently adjusted to the new experimental debhelper
compat level, but unfortunately commit da1b7a36 also implicitly
required debhelper 13. As a result, the test never ran. The point of
the test, however, is to select the experimental level when using the
version below. The test now runs as intended.

Signed-off-by: Chris Lamb <lamby@debian.org>