Calculate oldest possible age for test data in test preparation.
Based on HARNESS_EPOCH and file modification times, calculates the age
of a test specification. Used primarily for rebuilding decisions.

The resulting date is also used to fill templates (unless overridden).
Because of this secondary use, the value is further combined with
POLICY_EPOCH to make sure an old date does not trigger a timewarp
warning. Unless a test specification is more recent, the resulting
date is the oldest permitted according to policy. This may become
unnecessary once the disjointed tags proposal (aka tags.d) is
accepted.

Part of adaptive rebuilding for tests.

Signed-off-by: Chris Lamb <lamby@debian.org>