Make check_result() in test runner independent from test harness.
The function that checks the Lintian output took a harness dependent
test state as an argument. It also used a convoluted logic to make
sure that all Test-For tags were seen, and all Test-Against tags were
not. Rewrites the function to return an list of error instead, if
there are any.

Makes the test runner independent from the harness and may later aid
in the experimentation with other harnesses, such as TAP::Harness.

Signed-off-by: Chris Lamb <lamby@debian.org>