Use paths in test queue (instead of testcase data structure).
Use file paths to test specification directories to identify tests in
the run queue. It is more portable between harnesses and a
prerequisite to using TAP::Harness for the package tests.

Introduces a small change in the output format, but only for the
summary. For skipped tests, the suite is now part of the line for each
test.

Signed-off-by: Chris Lamb <lamby@debian.org>