Use relative paths as process labels instead of suite::testname.
The most portable way to refer to tests is by their paths.  In
addition, the concept of 'suite' for Lintian tests will change
considerably in the near future.  Changes the process label for each
test from ${suite}::${testname} to the relative path to the test
directory.

This changes the output format considerably and may take some time to
get used to. As a positive, the output very much resembles 'prove' or
TAP::Harness and may thus be familiar from other Perl contexts.

Signed-off-by: Chris Lamb <lamby@debian.org>