Run TAP scripts for all tests through TAP::Harness. (MR: !124)
Use TAP::Harness to run all Lintian tests. This unifies all tests. It
has the advantage that all results appear at the bottom. It is
therefore unlikely to miss a failure in the code quality scripts or in
the internal harness tests. Previously, those results tended to be
obscured by long output from the package tests.

The short output shown when building via 'debuild' (presumably for
lack of a tty) is no longer available. There may be a way to customize
TAP::Harness, if that is important.

Also, the command line setting $keep_going does not have any effect
currently. Somewhat surprisingly, such a "bail-on-fail" setting is not
available in Perl's standard TAP::Harness. It only requires a trivial
subclass available on CPAN, but was not used for now to avoid adding
another dependency. This can be addressed, if necessary.

As for speed, running all tests through TAP::Harness is currently
slower than the old runner. The difference is about 8 percent.  This
comes as a surprise since previous tests showed them about the same.
Perplexingly, the parallel test preparation seems responsible. It is
the only recent feature.

Other more modern harness implementations such as Test2::Harness offer
pre-loaded persistent runners. They could bring a speed gain that
should be investigated in time.

Signed-off-by: Chris Lamb <lamby@debian.org>