Move builder location for tests to 't/defaults/files'.
The builder file is a special template that is filled first and then
loaded into the template data. It allows other templates to receive
values seeded by the builder.

This mechanism is likely to be generalized in the future, so that
loading into the template data is possible from multiple files. Makes
the name of the builder file configurable.

Gbp-Dch: ignore