Remove declaration of default skeleton.
The default skeleton is 'default'. That was sufficient when skeletons
were specific to suites: Each suite had a default skeleton. When the
tag-related tests no longer have suites, those skeletons require more
meaninful names. The name of each such skeleton will then have to be
mentioned in any test specification that uses it. Removes the
declaration of a default skeleton.

Gbp-Dch: ignore

Signed-off-by: Chris Lamb <lamby@debian.org>