Remove suite as a consideration for skeleton selection.
The selection of skeletons was previously suite-specific. Remove all references
to suites from the test runner, except for the suites 'tags' and 'scripts'.
Also look for skeletons in a non-suite specific location. Also no longer store
the suite in the 'desc' file recorded in a test's work directory.

Gbp-Dch: ignore

Signed-off-by: Chris Lamb <lamby@debian.org>