Add a pedantic-only check to help detect files installed to /. (MR: !128)
As a way to help gather data and aid in future migration to /usr, as
well as a means of detecting packages that installed files to / to help
mount /usr that no longer need to, add an experimental pedantic-only
check for files in / (/bin, /sbin, /lib*) that are not compatibility
symlinks to the corresponding file in /usr.

Note that this is a pedantic-only check, with no plans to migrate it to
a warning or error in the near future and should not be used as the
basis for any form of mass bug filing.

Signed-off-by: Chris Lamb <lamby@debian.org>