Add default settings for how Lintian is called in tests.
This sets the default Lintian options in tests to '-I -E'. The new
option is intended to be used with the old 'Options': except that a
test using it no longer has to duplicate the default settings.  Makes
the test suite more internally consistent between tests. It is a
preparation for cycling through different output formats.

Gbp-Dch: ignore