Use new lintian invocation template when running lintian in tests.
Switches the runner from a programmatic invocation of Lintian to the
template-based call. This documents the options that were used and
their order, which makes debugging easier. It also aids in the
parameterization of the Lintian invocation. This is a preparation for
cycling through a variety of Lintian settings, such as output formats.

Gbp-Dch: ignore