Add a check for binaries using obsolete DES encryption. (Closes: #920376)
libcrypt.so.1 (part of glibc) used to provide a set of functions that allowed
raw use of the DES block cipher. Encryption with DES is now inherently insecure
(any use of single DES can be broken by brute force) and therefore these
functions were made inaccessible to newly linked programs in glibc 2.28.

Any program in the archive that unconditionally uses one of these functions
will FTBFS with glibc 2.28. Any program that conditionally uses one of these
functions may be falling back to an internal implementation of DES and
therefore have a security bug.

Therefore, I propose this new lintian check, which detects use of the functions
that were removed from libcrypt.so.1. In sid it is arguably redundant (since
any program that would fail this check will FTBFS anyway) but applying this
check to the stable archive would detect programs that were conditionally using
these functions and need to be audited for security bugs.

Signed-off-by: Chris Lamb <lamby@debian.org>