Adds tagdiff, which compares files produced by tagextract.
The tag specifications produced by tagextract have many advantages but
are designed to be readable by machines. This utility provides pretty
output for differences between two such files, hopefully even more so
in the future.

The output is sorted in reverse lexicographic order under LC_ALL=C.

Returns with a zero exit code even when the tags do not match. It
would be nice to match the behavior of standard 'diff', but that would
likely require disabling autodie and catching exceptions thrown by
Path::Tiny. For now, the exit code is non-zero only when the program
terminates abnormally.

When used in a terminal, produces colored output via Term::ANSIColor.
The test runner uses this tool to highlight discrepancies when tests
fail.

Gbp-Dch: ignore