Use output-invariant tag specifications with tagextract and tagdiff.
All expected tags in the test suite are now specified as EWI. This
makes it possible to calculate Test-For automatically. That will be
implemented in short order using a new desc parameter called Check.

The line ordering in 'tags' is from now on disregarded.

Makes tag specifications invariant of output format (selected via
Output-Format) and insensitive to order. The extracted tags are sorted
automatically.

With this change the test runner no longer compares output files
verbatim. Uses tagextract and tagdiff instead.