Remove internal Lintian warnings from tag specifications in tests.
Lintian emits some warnings that are not tags. Those warnings cannot
be processed by tagextract or tagdiff.

Please also see #921112, which tracks this issue.

The messages are emitted by lib/Lintian/ProcessablePool.pm and should
probably be converted to proper lintian warnings.

The following tests are affected:

t/tags/debs/control-field-traversal-4
t/tags/source/control-field-traversal-1
t/tags/source/control-field-traversal-3

The messsages relate to relative path names in control fields and look
like this:

    warning: tainted [...] package '...', skipping

The objective is to test for this warning. The tests do not contain
any other tags. As an interim solution, the tests were given the
ficticous tag 'package-tainted'. The test calibration mechanism is
used to remove the the internal warning from the actual output. This
causes the tests to fail. As a result, they were marked TODO.