Add an empty Test-Against to all tests that do not have one.
A new parameter called Check will make it possible to calculate
Test-For automatically. In terms of set theory, it is the intersection
of the tags belonging to the checks in the new parameter and the tags
expected in 'tags'.

A meaningful measure similar to Test-Against can likewise be
calculated. It is Test-For's complement, i.e.  the tags in the checks
that do not appear in 'tags'.

The mechanism works great for focused checks.  Unfortunately, many
checks are too broad. That could cause the automatic Test-Against to
contain too many tags. (The only drawback would be that more tests are
run when a tag is specified with 'onlyrun'.) Anyway, to ease the
transition, Test-Against is honored if present.

To preserve the old behavior even for test that do not have a
Test-Against, we add an empty field to them. To get the new
behavior, please simply remove it.

Test-Against should never be used in narrowly written checks.

Some tests are not tag-related. Test-Against is added only when
Test-For is present. The sole exception is runtests-todo-test-for,
which claims to test a ficticous tag 'lintian-easter-egg'. That test
will be repurposed to check if the computed Test-For (rather than the
actual one) is implausible.

Gpb-Dch: ignore