Set the new parameter 'Check' for all tests in suite 'tags'.
The purpose of the new parameter is to calculate Test-For
automatically. Adds the parameter with the appropriate value to all
tests in suite tags.

The singular in Check is aspirational. The field can contain multiple
checks separated by spaces.  The value is the the sum of all checks
belonging to the tags mentioned in Test-For and in Test-Against.

Gbp-Dch: ignore