Calculate Test-For and Test-Against from new parameter Check in runner.
A new parameter called Check makes it possible to calculate Test-For
automatically. In terms of set theory, it is the intersection of the
tags belonging to the checks in the new parameter and the tags
expected in 'tags'. Performs this calculation in the test runner.

The parameter Test-For is obsolete.

A meaningful measure similar to Test-Against can likewise be
calculated. It is Test-For's complement, i.e.  the tags in the checks
that do not appear in 'tags'.

The mechanism works great for focused checks.  Unfortunately, many
checks are too broad. That could cause the automatic Test-Against to
contain too many tags. (The only drawback would be that more tests are
run when a tag is specified with 'onlyrun'.) Anyway, to ease the
transition, Test-Against is honored if present.

To preserve the old behavior, an empty Test-Against is added for tests
that do not use it. To get the new behavior, please remove it.

Test-Against should not be used in narrowly written checks.