Repurpose runtests-todo-test-for for empty Test-Against.
This test checked for implausible tags in Test-For. With the parameter
Test-For obsolete, the test is repurposed to check for an implausible
computed Test-For (rather than an actual one).

When Test-Against is overridden, it cannot be empty while the expected
'tags' are also empty. That is because an empty Test-Against implies a
non-empty Test-For. That in turn requires something in 'tags'. Any
such test must fail.

To catch the failure, the test remains marked TODO.

This logic applies only to tag-related tests, which are those using
Check.

Gbp-Dch: ignore