Fix race; stop loop for collections after task fails. (Closes: 920691)
The primary issue was that the loop was not being stopped in case of a
failure, but the single additional line

+   $loop->loop_stop if not %{$running_jobs};

alone did not solve the issue. The way the processes were started may
have caused a race condition.

On that note, I am not sure that no race condition remains with
respect to @task. It may be better to open an IO::Async::Channel to
the parent and let the parent schedule the next task.

The commit following this one makes sure new tasks are also started in
a child in case of a failure.