Reassign watch-file-bug-765995 to check watch-file from changes-file.
The original test specification had

    Test-For: orig-tarball-missing-upstream-signature

so the test was erroneously assigned to the check changes-file.
Reassigning to watch-file.

Also, the bug report does not list a tag, but rather a Perl error
message. Remove Tag-Against. Configures this as a test against false
positives for any tags in check watch-file.

Gbp-Dch: ignore