Move runtests-empty-test-against to t/tags/tests; not check-related.
Tests a feature of the test runner whereby Test-Against cannot be
empty while 'tags' is also empty. (Actually, not even the subset of
'tags' belonging to Check may be empty.) This test is not tag-related
in the usual sense.  Moving back to t/tags/tests for further
evaluation.

The test needs a check, so 'manpages' cannot be removed. This test
will probably be transformed into script that checks this feature of
the runner via mocking.

Gbp-Dch: ignore