Tighten requirements for Test-Against vs expected tags.
The requirement that Test-Against may not be empty when 'tags' is
empty can be tightened. There could actually be tags present as long
as they do not belong to Check. In other words, the calculated
Test-For and the overridden Test-Against may not both be empty.

The behavior is being tested in runtests-empty-test-against from
commit ec81e0c9. We are not adding other valid tags to the test
because it is marked TODO. It would be too difficult to verify if the
test fails because a tag disappeared or because the calculation went
awry. The current situation is not perfect but more desirable.

Gbp-Dch: ignore