Demote the Test-Against test metadata field to be a selector for false-positives only. (MR: !142)
The field Test-Against used to have two functions. First, it would
help the harness determine which tests should be run when a tag was
selected. The second function was to generate errors when those tags
were present in the output.

The test suite calculates Test-Against automatically from Check and
'tags'. To preserve old behavior, Test-Against was allowed to override
the calculated value. As it turns out, that has dangerous side effects
when unrelated tags are ignored (a coming feature). The override
capability has been disabled.

Test-Against is now used only for the first purpose, which is to
select tests when running for a particular tag. Its only function is
to limit the tests that are run beyond those in which the tag does not
appear (but which belong to Check). Test-Against is thus just an
identifier for tests against false positives. Its use is expected to
fade, and eventually be obsolete, as broad checks become more narrow.

The danger of the override occurs when exact matching of tags is
turned off. This is a planned feature to reduce nuisance tags. Using a
new field Ignore-Unrelated-Tags, the test can discard any annoying
tags such as 'package-has-long-file-name' or, until recently, the
ubiquitious 'debian-watch-does-not-check-gpg-signature'. (It was
marked experimental.)

The test that checked this feature, 'runtests-empty-test-against',
will be discarded in the very near future.