Separate tests that track historical packaging; no truth value.
The two tests 'generic-dh-make-2005' and 'generic-dh-make-2008' exist
to track the performance of these historical file sets vis-a-vis the
latest version of Lintian. They come with specific instructions to
update the tags whenever necessary. Because of that, they do not have
much truth value. They should especially not count in calculations for
test coverage.  Separates them from the other tests.

The fields Check and Test-Against in the specifications for these
tests are meaningless. They are removed.

Gbp-Dch: ignore