Split up two tests for eleven checks each, and move the copies.
The tests that were each exactly associated with eleven checks each
were split and a copy was moved to each check. The following tests
were affected:

    t/tags/tests/legacy-foo++/
    t/tags/tests/legacy-debconf/

The fields for Test-Against were likewise split such that the tags
mentioned there belong only to the check with which that copy of the
test is associated.

Gbp-Dch: ignore