Split up two tests for fourteen checks each, and move them.
The tests that were each exactly associated with fourteen checks each
were split and a copy was moved to each check. The following tests
were affected:

    t/tags/tests/legacy-filenames/
    t/tags/tests/legacy-maintainer-scripts/

The fields for Test-Against were likewise split such that the tags
mentioned there belong only to the check with which that copy of the
test is associated.

Gbp-Dch: ignore