Split up two tests for fifteen checks each, and move them.
The tests that were each exactly associated with fifteen checks each
were split and a copy was moved to each check. The following tests
were affected:

    t/tags/tests/legacy-scripts/
    t/tags/tests/legacy-libbaz/

The fields for Test-Against were likewise split such that the tags
mentioned there belong only to the check with which that copy of the
test is associated.

Gbp-Dch: ignore