Split up test 'legacy-binary' for twenty checks, and move the copies.
One tests was associated with exactly twenty checks. It was split
and a copy was moved to each check. The following test was affected:

    t/tags/tests/legacy-binary/

The fields for Test-Against were likewise split such that the tags
mentioned there belong only to the check with which that copy of the
test is associated.

Gbp-Dch: ignore