Split up six tests for three checks each, and move them.
The tests that were each exactly associated with fourteen checks each
were split and a copy was moved to each check. The following tests
were affected:

    t/tags/tests/control-file-rules-requires-root-no/
    t/tags/tests/binaries-multiarch/
    t/tags/tests/control-file-rules-requires-root-binary-targets/
    t/tags/tests/binaries-missing-lfs/
    t/tags/tests/changelog-file-stable/
    t/tags/tests/cruft-general-wig-pen/

The fields for Test-Against were likewise split such that the tags
mentioned there belong only to the check with which that copy of the
test is associated.

Gbp-Dch: ignore