Split up sixty tests for two checks each, and move them.
Sixty tests that were each exactly associated with two checks each
were split and a copy was moved to each check.

The fields for Test-Against were likewise split such that the tags
mentioned there belong only to the check with which that copy of the
test is associated.

Gbp-Dch: ignore