New make-based builder for binary Debian packages.
Replaces the t/bin/create-deb that was previously used for most tests
with skeleton 'deb'. That program was not usable for all tests. The
current solution no longer depends on that program.

This builder has not been tested well with non-native packages (there
is only one test with this skeleton that uses it) and will probably
fail. The t/bin/create-deb program is kept in hope it may provide
guidance when such functionality is being provided.

Gbp-Dch: ignore