Add test selector 'skeleton:' to effectively select the old test suites.
Some work remains to be done with templates for the old suites. The
skeletons correspond more or less to those suites. This change makes
it easier to test a template, as all affected tests run at the same
time.

The goal is to devolve Makefiles that currently create static files
into files in the filesystem. Another goal is to reduce the number of
customized Makefiles, whenever a test diverges from the template.

Gbp-Dch: ignore