In tests, save tag differences in file 'tagdiff' in working directory.
Saves the tag difference in a file in the test working directory. In
the near future, that will enable the interactive adjustment of tag
specifications. It will be easier to recalibrate tests after changes
were made to Lintian checks. There will be no more cut and paste.

In all fairness, manual copying of tags is also no longer an option.
The tag differences are currently kept in a format that is different
from the file 'tags'. (We may convert tags at some point.) Some
differences in format are: the initial letter code is missing, and the
package type is always displayed.

Gbp-Dch: ignore