Add utility 'tagadjust' to automatically modify tag files in tests.
This small utility makes it possible to recalibrate the test suite
after changes were made to Lintian checks. It is no longer possible to
copy differences manually because the formats are different. One can
use this program separately.

Allows the interactive selection of adjustments when the command line
switch '-i' is used.

Gbp-Dch: ignore