Split runtests processing between scripts and tag tests.
The 'scripts' and tag tests are still processed in sequence, but they
run separately through the same TAP::Parser::Aggregator. This makes it
possible to interactively adjust the tag tests, for example after
changes were made to Lintian checks. In the future, we will refer to
this action as calibration of the test set.

Terminates a test run if the scripts fail unless $keep_going is set.
That prevents the whole suite from running when there may be more
basic problems with setup.

Gpb-Dch: ignore