Show untested tags in test runs.
This test script checks if all tags known to Lintian are also seen in
the test suite. The test suite is presently incomplete, and some tags
are missing. Those tests are automatically marked TODO. For now, the
missing tags will not cause the test run to fail, but are announced as
untested.

Once the test suite has full tag coverage, the TODO classification
should be removed. Then missing tags will trigger a build error.