Use the new universal tags, similar to EWI, in test runner.
Changes the name suffices for generated files to make sure they do not
conflict with older ones in the old format. It may be a good idea to
execute 'rm -rf debian/test-out' in response to this commit, but it is
not necessary.

Also reverses the argument order for check_result().

Gbp-Dch: ignore