Change runtests options and arguments for stand-alone operation.
The script t/bin/runtests is customarily invoked via debian/rules
because it ensures the existence of the test working directory, among
other things. On a day-to-day basis, however, it will soon be more
convenient to invoke it directly as t/bin/runtests.

The invocation in debian/rules will suppress any prompts as may arise
when the test set is calibrated. It is enabled by default and thus
only accessible when using t/bin/runtests. This commit changes many
options and argument to make such a stand-alone operation super easy.

Most important, in the future no command line options are required.
The testset defaults to 't'. The test output directory defaults to
debian/test-out in the same parent directory as the testset. The test
output directory is created unless it exists.

The onlyrun option, which is used to select tests, is now a proper
option and no longer a simple argument. It has to be prefaced with
'--' unless the shorthand -o is used.

Please note that the onlyrun option to *debian/rules* does not work
any more. It is unclear if it is still needed.

Also sets up an option for unattended use, which rules out stopping
for answers as might be asked when calibrating the test suite.  This
option will be implemented when the interactive calibration is
introduced.