Interactively calibrate the expected tags in failing tests. (MR: !145)
When a terminal is detected and the command line option for unattended
operation is not enabled, offer to re-calibrate the tags expected in
tests that had errors.

After adding new features in response to a bug report, it would now be
easy to run:

    t/bin/runtests --onlyrun=check:XXX

The command will ensure that the edits were successful and offer to
recalibrate any tests in which the new tag, if any, might appear.

Signed-off-by: Chris Lamb <lamby@debian.org>