Fix a serious bug in get_tagnames() in T::L::Output::Universal.
This is a vital routine to ensure good test results. Fortunately, it
was not used exclusively to pass tests. In fact, no failing test could
pass a result of the bug. Also used in test selection. Fixes the bug.

An internal harness test for this function is forthcoming.

Gbp-Dch: ignore