Fix post-test sed script for files-old-config-script.
The substitution in this post-test sed script forgot to add space back
in for some matches. For some reason, it was not noticed when the
expected tags were formulated. The tag specification included tags
commingled with their informational messages (in this case, a path).
That, in turn, made it look like the test emitted tags unknown to
Lintian.

Adds the space and adjusts the expected tags so that the space
appears. The tags file is now parsed correctly.

Gbp-Dch: ignore