Add post_build scripting capability to make-based builder for tests.
The missing tag checksum-count-mismatch-in-changes-file is most easily
implemented by changing the *.changes file after the upload was built.
Adds this capacity to the make-based builder.

Some tests from the old changes suite may be reimplemented with the
same technique.

Also moves the similar pre_build script to the same template
whitelist.

Gbp-Dch: ignore