Update recommendation in an update-menus long description that will not trigger a check for full path names in maintainer scripts. (MR: !149)
The tag maintainer-script-does-not-check-for-existence-of-updatemenus
recommends a solution which triggers
command-with-path-in-maintainer-script. That tag then recommends a
better solution. This change applies the recommendation from the latter tag
description to the text for the former as a convenience for users.

The issue is visible in the test menus::executable-menu-file, which
will be adjusted in the future.

Signed-off-by: Chris Lamb <lamby@debian.org>