Fix detection of security features in Portable Executable (PE) binaries.
The Portable Executable (PE) check did not come with any true positive
tests (only two false positives) and did not appear to work on PE32 or
PE32+ binaries generated with MinGW-CC. The previous version used an
incorrect offset into the object table (90 != 94). The logic around
the 'any' check was also reversed. This commit works with the tests
provided in additional commits.

The tests that follow use the 'genpeimg' tool from MinGW to lower the
security settings on a 'safe' binary that was generated with the
compiler's default settings.

Also adds a list of the missing security features to the tag. They are
at the end, after the file name.