Capture output from intltool-update and show as N: messages instead of plain print.
The tag parsing tool tagextract transforms both the expected tags and
the actual tags into a universal format for comparison. Discarding
input lines is dangerous, as two empty files will always match.  An
error may always produce positive test results.

For that reason, tagextract discards only lines starting with 'N:'. It
fails on other, unexpected input. This commit converts the direct
messages printed by intltool-update into messages starting with N:.
While the information is still printed for the user, the new format
makes it safe to discard those lines when comparing tags.

This behavior may change if the expected tags are converted into the
universal format on disk in the future. In that case, the expected
tags would never be modified and always stand on their own. The runner
would also be faster. The matter is under consideration.

Gbp-Dch: ignore