Use green and red terminal colors when adding or removing tags during calibration. (MR: !153)
Justifies all tags to the left margin. That makes it easier to compare
tags visually when removing one and adding another, especially if the
changes are minor and the strings similar.

Clearly separates the tag content from the action prompt. The colors
are a bit in your face but they catch a user's attention, which is
probably a good idea for a calibration task. Minimizes the chance for
confusion or robotic acceptance of the proposed changes.