Ensure unpack loop stops when process succeeds after processable failed to try and prevent Lintian from occasionally hanging. (Closes: #922737)
This item was left over from commit 6f063218. I questioned the return
statement at the time but felt discouraged by the explicit comment
instructing not to break the loop.

The issue is that there is a rare situation, which could correlate
with the incident rate of 1 per 2000 reported in #922737, that the
loop is not stopped if a process succeeds while $procid already failed
elsewhere.  (Here, $procid refers to a "processable", also known as a
Lab::Entry, and not the PID.)

If this does not resolve the bug, the next step is to rely more
consistently on IO::Async. Based on conversations on #perl-help, it is
not recommended to mix IO::Async with other process management
techniques such as IPC::Run (and probably even waitpid) that are used
in Lintian. That would be a somewhat larger undertaking.

This change does not address the potential remaining race condition
mentioned in commit 6fe7c45e, which related to the shared data
structure of tasks outstanding.

Signed-off-by: Chris Lamb <lamby@debian.org>