Replace call to 'rm -rf' with Path::Tiny->remove_tree in Lab::Entry::remove_async.
This is a further attempt to identify the thread that gave rise to
Bug#922737. The thread was labeled 'rm'. This sets the process label
to something related to Lab::Entry removal on platforms that support
it.

Also adjusts the logic around $status. When using a 'code' block (vs.
command) is IO::Async::Loop->spawn_child, $dollarbang is always set to
something.

Gbp-Dch: ignore