Explain in description for file-references-package-build-path that Build-Path is optional in *.buildinfo
Lintian will not emit this tag unless the *.buildinfo includes the
field 'Build-Path'. The field is optional. In most cases, one must set
DEB_BUILD_OPTIONS=buildinfo=+path or use the option
'--buildinfo-option=--always-include-path' with dpkg-buildpackage when
building. Mentions this in the tag description.

Also adds a reference. That fields does not appear to permit
continuation lines.

Gbp-Dch: ignore