Add library modules with tag parsing routines for the different output formats.
Move the tag parsing routines for the various output formats from
tagadjust to modules in the library path. It will allow to keep the
data in memory without temporary files.

Having these routines available is expected to speed up tag coverage
calculations, which currently involve a lot of conversions via
tagextract from the EWI format of the on-disk tag specifications to
the universal format used for interactive calibration.

At some point, the on-disk format for tags may change to the universal
format.

Gbp-Dch: ignore